Support internal metric rendering in Prometheus text format in C++
Maps all internal metrics to one or more labelled time series.
Due to poor compatibility between the data model (and sampling
strategy) of the legacy metrics framework and that of Prometheus,
all time series are emitted as `untyped` metrics.

This is a stop-gap solution on the way to "properly" supporting
Prometheus exposition, and the output of this renderer should
therefore only be used for internal purposes.