Add SimpleTokenScript to SimpleTokenizer
When parsing datasets such as WikiDumps to a significance model, we want
to only keep characters of that language script within our model. So
when adding the script value to our tokenizer we are able to use this to
filter out non-latin words when creating an english significance model
for example.