Support pipelining (batching) of mutating ops to same bucket
Bucket operations require either exclusive (single writer) or
shared (multiple readers) access. Prior to this commit, this
means that many enqueued feed operations to the same bucket
introduce pipeline stalls due to each operation having to wait
for all prior operations to the bucket to complete entirely
(including fsync of WAL append). This is a likely scenario when
feeding a document set that was previously acquired through
visiting, as such documents will inherently be output in
bucket-order.

With this commit, a configurable number of feed operations
(put, remove and update) bound for the exact same bucket may
be sent asynchronously to the persistence provider in the
context of the _same_ write lock. This mirrors how merge
operations work for puts and removes.

Batching is fairly conservative, and will _not_ batch across
further messages when any of the following holds:

 * A non-feed operation is encountered
 * More than one mutating operation is encountered for the
   same document ID
 * No more persistence throttler tokens can be acquired
 * Max batch size has been reached

Updating the bucket DB, assigning bucket info and sending
replies is deferred until _all_ batched operations complete.

Max batch size is (re-)configurable live and defaults to a
batch size of 1, which shall have the exact same semantics as
the legacy behavior.

Additionally, clock sampling for persistence threads have been
abstracted away to allow for mocking in tests (no need for sleep!).