Use better non-strict cost estimates for btree and disk index iterators.
We previously increased the strict cost of heap-based complex blueprints (e.g OR)
to avoid the problem of them being forced strict too early.
More benchmarking however has shown that its better address the real problem,
which is the non-strict cost of btree iterators that are childen of the ORs.
Its important that the non-strict cost is most accurate with high inflows (moving towards 1.0),
as the time penalty of choosing wrong is at its highest then.

This adds additional benchmarks needed for analysis and
updates/adjusts the parameters and formulas in flow_tuning.h to match observations.