Remove legacy storage node root directory IO in unit tests
Once upon a time, VDS roamed the lands. It used real disk IO as
part of tests. Then came the meteor and in-memory dummy persistence
took over. Now it is time for the fossils to be moved into a museum
where they belong.

Also make PID file writing conditional on a config that is set to
`false` during unit testing (but `true` as default).