Make replica selection for Puts and bucket activation symmetric
The legacy Put replica selection behavior may route new versions
of a document to replicas that are not considered optimal for
activation. This is not normally an issue, but can manifest itself
as missing coverage when the system is in flux with replicas
moving away from Retired nodes containing ready replicas, as the
existing replicas on the Retired node would be preferred for
activation (and thus be used for searches) but incoming Puts would
instead be sent to non-retired nodes due to being in the ideal state.

The new replica ordering (and transitively, selection) behavior is
identical between Puts and activation. This should help ensure
that new versions of the document is routed to replica(s) that
are most likely to be visible as part of searches.

New selection behavior for Puts is config-gated and defaults to
the legacy behavior.

This also subtly changes the fallback ordering criteria for replica
activation to consider the replica's existing DB _entry_ order
instead of its node index. Since DB entries are always ordered by
their ideal state order (with Retired nodes included), this will
evenly distribute fallback activations rather than skewing them
towards lower indexes. It is not expected that this has any negative
effects in practice, and is therefore _not_ a config-gated change.