Ensure that only 1 operation per document ID can be in flight per diff apply
It's possible for a diff to contain multiple versions of the same
document, persisted at different timestamps. To avoid async scheduling
multiple operations per distinct document ID (technically a violation
of the SPI invariants), do a pre-pass of the diff to find the highest
timestamp present in the diff per document. Only schedule an operation
if it's the newest one for a document.