Add metric for number of nodes not converged to the latest cluster state version
There is a grace period (currently hard-coded to 30s) between the
time a new state version is published and before nodes are counted
as not converging. This state is not sticky for a given node, so
if the state keeps changing constantly it's possible that we under-
count the number of non-converging nodes. But this can be improved
over time; for now it's primarily a source from which we can generate
alerts, as failures to converge can manifest itself as other failures
in the content cluster.

Also move out of sync ratio computations to the periodic metric
update functionality and reset it if the current controller is not
the believed cluster leader.

Start moving some CC timer code to use `Instant` instead `long`s with
implicit units.