Add distribution config bundle to `ClusterStateBundle` in C++
Move distribution config transforms to vdslib so that the distribution
config bundle can contain derived configs for all bucket spaces in
one central place.

This is part of the prerequisite work needed before we can start
pushing distribution config from the cluster controller and rewiring
how distribution config is propagated and used in the backends.

Also, rename `Distribution::serialize()` to `Distribution::serialized()`
since it returns a const ref to a cached serialized form and does not do
on-demand serialization.