Speed up insertion into HNSW index by referencing vector memory in more cases.
When inserting vectors into a HnswIndex we use DistanceFunctionFactory::for_insertion_vector()
to create a distance function each time distance calculations are needed for a candidate vector.
All these vectors are of the same type, as given by the tensor type of the TensorAttribute.
The lifetime of vectors is also handled by them either being stored in the TensorAttribute,
or existing in the document being inserted.

This PR speeds up insertion for vectors of type int8, float and double.
This is done by referencing the vector memory directly in the distance function,
instead of copying (and transforming) into a tempory memory buffer.
Vectors of type bfloat16 are still transformed to float before distance calculations.