Add logging functionality using `std::format`
This adds a separate `logf.h` header (which transitively includes
`log.h`) with a `LOGF` macro that supplements the existing `printf`
syntax-based `LOG` macro.

Usage is identical to the existing macro, except that `std::format`
semantics are used for the formatting string and arguments (which
means no more `c_str()`) and the string is checked for validity at
compile-time.

Example:

  LOGF(debug, "Have fed {} kittens", 42);

For now, the code is conditionally compiled if `__cpp_lib_format`
is defined, which means it cannot be used until this is the case
for all compilers we use.