Collapse redundant document selection result set entries in Java
AND/OR operators have an evaluation short circuit for `false` and
`true` respectively, but will otherwise fallback to an expensive
matching loop in order to support (deprecated) variable matching
semantics. The expected case is that all variable maps are empty,
in which case we have no need for more than a single result set
entry for any given `Result` enum value.

This reduces expected result set _size_ from O(|lhs|*|rhs|) to O(1),
though the _time_ complexity is still O(|lhs|*|rhs|).

This should be functionally identical to an existing optimization
done for the exact same scenario in the C++ implementation. It is
possible to optimize certain cases much further, but that would
require less conservative changes.