Remove cluster state version downgrade fallback
To avoid inherent race conditions with overlapping cluster
controller leader intervals (caused by the old leader not yet knowing
it has been deposed) where both an old state version and a
newer state version is concurrently published, we want to only
accept strictly increasing version numbers (for the lifetime of
a process; these are currently not durably stored on content nodes).
These version numbers are backed by a ZooKeeper quorum, ensuring that
they _are_ durably stored for cluster controllers.

A content node only observing strictly increasing version numbers is
an invariant that holds _unless_ an explicit fallback is triggered,
where we can still accept an older version.

This fallback was intended to be a "failsafe" if ZooKeeper state on
the cluster controllers was lost, but its implementation depended
on information that is not actually present in all CC RPCs, meaning
that it could kick in even when not intended, thus rendering the
race condition protection void.

The CC RPC in question is not easily extensible, so instead remove
the fallback entirely. This has the bonus of content nodes actually
being able to rely on the version invariant internally. Downside is
that content node and distributor processes must be restarted to
accept a lower state version upon ZK state loss, but in that case
you probably have bigger problems.