Use timestamp TaS conditions from Document V1 handler when possible
Timestamp-predicated test-and-set operations are cheaper to evaluate
in the backend, are safe in the face of concurrent feeding and work
as expected when the batch selection uses imported fields (since the
selection itself is only evaluated during the visiting phase, _not_
during the mutating operation processing phase).

Since we now receive backend persistence timestamps as part of visiting,
patch these into the generated `TestAndSetCondition` instances sent
for each resulting update/remove operation.

To allow for backwards compatibility, the generated TaS condition
has _both_ a required timestamp as well as the original batch operation
selection. The selection is only used as a fallback if backends are
on a version too old to support timestamp requirements. On nodes with
timestamp TaS support, the selection is _ignored_ entirely when a
required timestamp is present.

Also note that if a content node is too old to support sending
persistence timestamps, it will default to 0 the receiver which also
preserves legacy behavior.