Ensure TaS fieldset enumeration descends below binary arithmetic AST node
This resolves a long-standing (but not discovered before now) bug
where the AST visitor responsible for determining the minimal set
of fields to fetch from the backend in order to evaluate a test-and-set
condition did not descend into the children of a binary arithmetic
operator node. This meant that if an expression only mentioned a
particular field once and this happened as part of a binary arithmetic
sub-expression, the field would not be fetched and the condition
evaluation would fail.

The fact that no one has observed this thus far means that this seems
to be a very rare use-case in practice...