Add a Count-Min-based probabilistic relative frequency sketch
Adds an implementation of a probabilistic frequency sketch that allows
for estimating the relative frequency of of elements from a stream of events.
That is, the sketch does not capture the _absolute_ frequency of a given
element over time.

To reduce the requirement for the number of bits used for the sketch's
underlying counters, this sketch uses automatic decaying of counter values
once the number of recorded samples reaches a certain point (relative to
the sketch's size). Decaying divides all counters by 2.

The underlying data structure is a Count-Min sketch [0][1] with automatic
decaying of counters based on TinyLFU [2].

This implementation has certain changes from a "textbook" CM sketch,
inspired by the approach used in [3]. In particular, instead of having `d`
logical rows each with width `w` that are accessed with hash-derived
indexes (and thus likely triggering `d` cache misses for large values
of `w`) we subdivide into w/64 blocks each with fixed number d=4 rows of
32 4-bit counters, i.e. each block is exactly 64 bytes. Counter updates or
reads always happen within the scope of a single block. We also ensure the
block array is allocated with at least a 64-byte alignment. This ensures
that a given sketch update will touch exactly 1 cache line of the
underlying sketch buffer (not counting cache lines occupied by the sketch
object itself, as we assume these are already present in the cache).
Similarly, comparing the frequency of two elements will always touch at
most 2 cache lines.

The Count-Min sketch (and its cousin, the Counting Bloom Filter) using `k`
counters is usually described as requiring `k` pairwise independent hash
functions. This implementation assumes this requirement is unnecessary
assuming a hash function with good entropy; we instead extract
non-overlapping subsets of bits of a single hash value and use these as
indices into our data structure components.

References:
 [0]: The Count-Min Sketch and its Applications (2003)
 [1]: https://en.wikipedia.org/wiki/Count%E2%80%93min_sketch
 [2]: TinyLFU: A Highly Efficient Cache Admission Policy (2015)
 [3]: https://github.com/ben-manes/caffeine/blob/master/caffeine/
      src/main/java/com/github/benmanes/caffeine/cache/FrequencySketch.java