Add explicit trimming support to `lrucache_map`
Trims the cache size so that it is within its capacity limits.
Since the cache itself will normally do this during inserts,
this can be used to explicitly trim the cache when higher-level
capacities (used via `removeOldest()`) change, which are not
picked up directly by the cache itself.

Note that this does not use soft limits; if the cache has only
a single element, and it is over-sized, trimming will remove
the entry.

Unify trimming and `removeOld` into a single function.

Also make `verifyInternals()` work with an empty mapping.