Integrate optional LFU frequency sketch into (S)LRU cache
Allows a cache to be configured to use a probabilistic LFU frequency
sketch for gating insertions to its segments. An element that is a
candidate for insertion will only be allowed into a segment if it is
estimated to be more frequently accessed than the element it would
displace. LFU gating works in both LRU and SLRU modes. In both modes,
initial insertion into the probationary segment is gated (for cache
read-through and write-through). In SLRU mode, promotion from
probationary to protected is also gated. In the case that promotion
is denied, the candidate element is placed at the LRU head of the
probationary segment instead, giving it another chance.

The configured sketch element count should be at least as large as the
maximum _expected_ number of elements that the cache can hold at once.

The default size is 0, i.e. LFU functionality is disabled.

Setting the size to >0 will always create a new sketch. The sketch
will be initialized with the cache keys that are currently present
in the cache segments, giving each existing entry an estimated
frequency of 1. All preexisting frequency information about entries
_not_ currently in the cache will be lost.