Ensure document is visible in render queue prior to visitor ACK
Document V1 API visitors asynchronously invoke a callback every
time a document is received from a content node. This callback
previously ACKed the operation, rendered the document JSON to a
byte buffer and enqueued the buffer to an internal render queue
(in a lock-free manner).

This ordering has a subtle flaw. If an ACK is sent for the final
document of a visitor, it is possible (but unlikely) that the
content node can observe the completion of the whole visitor and
unwind this entirely back to the client (via the distributor)
_before_ the document JSON rendering and buffer enqueueing has
had time to complete. This will invoke another completion callback
which flushes all remaining document buffers to the wire and marks
the response as done. Any buffers not on the queue at that point
will never be rendered, i.e. will appear to the client as if lost.
This is what happens in the above ACK race.

This commit simply moves response ACKing to explicitly happen-after
JSON rendering and buffer enqueueing to avoid premature session
completion.