Add syntax for filter-threshold for a rank profile.
A number between 0 and 1 that specifies the threshold value
for when a query term is considered a filter in backend matching,
and cheaper bitvector posting lists can be used.
Is triggered when the hit estimate of the query term is larger than the threshold value.

Note: This functionality is not yet wired in the backend,
so only the schema syntax is available for now.