Reduce max relative merge memory usage from 3% to 1.5%
Merge memory limits are by default set as a percentage of the
process memory limit, capped by lower and upper bounds.

On nodes with the vast majority of memory taken up by attributes,
using 3% of memory on merging might be a much larger percentage
jump in _heap_ memory, which can cause feed block limits to be
crossed if the node's already operating very close to the limit.