Validate no overlap of field rank `filter` and `filter-threshold` declarations
A field ranking declaration cannot both state that a field should
have `rank: filter` and _also_ have an explicit `filter-threshold`.
This applies to both whether rank filter is declared in the document
schema or in the rank profile (`filter-threshold` can only be
declared in the rank profile). This restriction also holds if
the rank profile is inherited.