Implicitly bump priority of `searchvisitor` instances
When the Document API protocol serialization was rewritten to
use Protobuf, the legacy client operation priorities were
intentionally omitted and instead all default to the same value.
This is because operation priorities for mutations cause a
lot of headaches on the backend nodes due to potential reordering
in the persistence queues.

However, this introduced a subtle regression where certain
read-only operations that _should_ have a high priority no longer
had them, in particular streaming search.

As a pragmatic solution, this commit adds an implicit priority
bump of visitors of type `searchvisitor`, which is the visitor
instance type used by streaming search.