Do not trigger window resize if time delta is zero
Prior to this commit, if the delta time in milliseconds between two
candidate window resize events was zero, the local throughput would
be computed as a floating point division of zero. I.e. positive
infinity. This is likely somewhat optimistic when compared to the
_actual_ throughput on the system. This would also always trigger
an attempt at a window size increase.

Simply refrain from adjusting the window size based on throughput
until there is a non-zero time delta.