Use same priority for bucket size split/join as for default feed
Having the same priority as feed maintains FIFO ordering across
nodes and removes the possibility for any lingering, subtle
priority inversion cases.

The previous priority values were introduced when splitting and
joining buckets were very expensive operations that required a
lot of IO. That is not the case in today's implementation.