Add configurable max document mutation operation message size
Iff set to a value > 0, the distributor will enforce a size
check on Put and Update operations arriving over the Document API,
where over-sized messages will be explicitly rejected with a
non-retryable error code.

Intended as a safety measure to prevent overly large documents
from being accepted into the cluster.

The message size is estimated based on the uncompressed on-wire
serialized payload size, so it's only a (possibly inaccurate)
proxy for the actual impact the underlying document operation
will have on the backend. But it's a "free" heuristic since we
already have that information available from the RPC subsystem.

Any number <= 0 implicitly defaults to 2 GiB (INT32_MAX),
i.e. effectively unbounded.