Aggregate distributor-provided error reports on the cluster controller
This adds parsing and handling of the `response-stats` field in the
HostInfo payload from distributor nodes. Currently only network errors
are considered. These are further aggregated per distributor per
content node, similar to how we track bucket space-specific statistics
already. Aggregation is sparse, so in a healthy cluster the expected
additional overhead is minimal.

Unlike the existing statistics, error statistics are updated even if
the distributor has _not_ converged to the most recent cluster state
version. This is to avoid catch-22 situations where a distributor fails
to converge to the newest version due to connectivity errors, but
where error statistics are not tracked and reported because the node
has not yet managed to converge.

In this initial implementation, errors are only surfaced on the cluster
controller status page as part of the content node information view.
If at least one distributor reports network errors, a warning will
be printed for the node. If the reporting distributor is also not
converging to the latest cluster state version, an explicit convergence
warning will also be printed.

Aggregation is currently gated by a new live config parameter. If
disabled, existing aggregated information will be cleared.