Improve and extend autovectorization target selection
Main new features:
 - A new AVX3_DL (~Icelake Server) target for x86-64.
 - Runtime detection of CPU support on aarch64.
 - Fallback to generic NEON on aarch64 when no fp16+dotprod
   instruction extensions are present

The aarch64 (ARM64) CPU detection is currently only used to
_downgrade_ the acceleration kernels relative to the default
build options. This is required to get Valgrind to work, which
only supports the baseline NEON instructions.

Unify naming convention of existing AVX-512 autovectorized
target to `AVX3` to match Google Highway. Since this targets
Skylake, a few extra CPU constraints have been added to the
existing AVX3 selection logic to ensure all required Skylake
AVX-512 extensions are present. This is not expected to change
the outcome of the accelerator selection in practice.

The new AVX3_DL target is currently _not_ enabled by default;
I want to test it in a controlled manner in our CI first.
To allow for doing this, an internal environment variable has
been added, which lets the desired CPU target level be specified.
The minimum of the requested and supported CPU level for a
given platform is then used to select the runtime implementation.

Based on testing on a Sapphire Rapids (which is the Icelake
successor) system, performance numbers between AVX3 and AVX3_DL
are largely the same, except for popcount, where AVX3_DL
performance is roughly _doubled_, presumably due to using
`VPOPCNTDQ` instructions.