Defer `int8` (`Int8Float`) dot product calculation to vectorized kernel
Since `Int8Float` is just a truncated version of the integer part of
a `float` there's no exponent/mantissa distinction to worry about, i.e.
we can cheekily reinterpret the whole thing as an array of `int8_t`
and pass it on to the vectorized implementation. This is effectively
identical to what's already done for vectorized tensor L2 distance evals.

Since we now have CPU feature detection and target selection on aarch64,
this will work with Valgrind even without NEON `dotproduct` support.