Only log lidspace compaction (re-)start events when debug logging is enabled
lidspace compaction throttling might cause a lot of stop-and-go
for compaction, causing an excessive amount of (re-)start edge
events. Should find a more elegant solution for this, but for now
simply avoid logging any events for these cases unless `debug`
level is active for the event logger component. Completion events
are unconditionally logged as before.