Add resource limiting to `SharedOperationThrottler`
This adds a second configurable dimension to the throttling
behavior, where token acquisitions can specify an underlying
resource usage (in an intentionally unspecified numerical unit)
and the throttler enforces that the total amount of resources
used is within the configured limit. If the resource usage would
be exceeded, token acquisitions will block (or fail, if using
the polling API) as-if the underlying throttle policy had refused
further tokens.

To ensure liveness, at least 1 token will always be allowed to
be acquired, even if it exceeds the configured limit. In other
words, this is a soft limit.

As with other throttler parameters, the resource limit can be
configured live. By default, resource limiting is _not_ enforced.