Ensure well-defined global distribution over heterogenous group topologies
Prior to this commit, global distribution was implemented by
"rewriting" a non-global distribution config into one where the
redundancy/ready copies was equal to the total leaf node count,
and where the partition specs should evenly distribute replicas
across sub-groups.

This works as expected when all groups have the same size, i.e.
the topology is homogenous. When this is _not_ the case this
approach would subtly break down due to the non-obvious semantics
of partition specs, where the assignment of per-group redundancies
given by the spec is _independent_ of the ordering of the groups
in the config. This means that for any given bucket, two groups of
5 and 3 nodes could end up with a redundancy computation of 3 and 5,
respectively (and vice versa, in which case the problem would be
effectively invisible). This would lead to missing replica
placement for a subset of the buckets.

With this commit, global distribution is now an explicit parameter
to the underlying `Distribution` object. Iff set, the partition
spec is completely ignored when computing group redundancies, and
the effective redundancy for any group is always set to be the
number of descendent nodes in, or under, that group.

However, we still have to perform the legacy distribution config
rewrite since its output is used as part of the distribution config
"hash" that is communicated from distributors to content nodes as
part of bucket info handshakes. If these mismatch across versions,
convergence will fail.